requirements.txt
psycopg2-binary==2.9.3
celery==5.4.0